+++ internal/scaffold/demo_test.go
@@ -2,7 +2,11 @@ package scaffold
 
 import (
 	"fmt"
+	"os"
+	"os/exec"
 	"testing"
+
+	"github.com/rs/zerolog/log"
 
 	"github.com/httprunner/hrp"
 )
@@ -11,6 +15,21 @@ var (
 	demoTestCaseJSONPath = "../../examples/demo.json"
 	demoTestCaseYAMLPath = "../../examples/demo.yaml"
 )
+
+func buildHashicorpPlugin() {
+	log.Info().Msg("[init] build hashicorp go plugin")
+	cmd := exec.Command("go", "build",
+		"-o", "../../examples/debugtalk.bin",
+		"../../examples/plugin/hashicorp.go", "../../examples/plugin/debugtalk.go")
+	if err := cmd.Run(); err != nil {
+		panic(err)
+	}
+}
+
+func removeHashicorpPlugin() {
+	log.Info().Msg("[teardown] remove hashicorp plugin")
+	os.Remove("../../examples/debugtalk.bin")
+}
 
 func TestGenDemoTestCase(t *testing.T) {
 	tCase, _ := demoTestCase.ToTCase()
@@ -25,6 +44,10 @@ func TestGenDemoTestCase(t *testing.T) {
 }
 
 func Example_demo() {
+	buildHashicorpPlugin()
+	defer removeHashicorpPlugin()
+
+	demoTestCase.Config.ToStruct().Path = "../../examples/debugtalk.bin"
 	err := hrp.NewRunner(nil).Run(demoTestCase) // hrp.Run(demoTestCase)
 	fmt.Println(err)
 	// Output:
@@ -32,6 +55,9 @@ func Example_demo() {
 }
 
 func Example_jsonDemo() {
+	buildHashicorpPlugin()
+	defer removeHashicorpPlugin()
+
 	testCase := &hrp.TestCasePath{Path: demoTestCaseJSONPath}
 	err := hrp.NewRunner(nil).Run(testCase) // hrp.Run(testCase)
 	fmt.Println(err)
@@ -40,6 +66,9 @@ func Example_jsonDemo() {
 }
 
 func Example_yamlDemo() {
+	buildHashicorpPlugin()
+	defer removeHashicorpPlugin()
+
 	testCase := &hrp.TestCasePath{Path: demoTestCaseYAMLPath}
 	err := hrp.NewRunner(nil).Run(testCase) // hrp.Run(testCase)
 	fmt.Println(err)