+++ runner.go
@@ -9,8 +9,11 @@ import (
 	"net/http"
 	"net/http/httputil"
 	"net/url"
+	"os"
+	"os/signal"
 	"strconv"
 	"strings"
+	"syscall"
 	"testing"
 	"time"
 
@@ -19,6 +22,7 @@ import (
 	"github.com/rs/zerolog/log"
 
 	"github.com/httprunner/hrp/internal/ga"
+	"github.com/httprunner/hrp/plugin/common"
 )
 
 // Run starts to run API test with default configs.
@@ -155,12 +159,19 @@ func (r *caseRunner) reset() *caseRunner {
 }
 
 func (r *caseRunner) run() error {
+	config := r.TestCase.Config
+
+	// init plugin
+	var err error
+	if r.parser.plugin, err = initPlugin(config.ToStruct().Path); err != nil {
+		return err
+	}
 	defer func() {
 		if r.parser.plugin != nil {
-			r.parser.plugin.quit()
+			r.parser.plugin.Quit()
 		}
 	}()
-	config := r.TestCase.Config
+
 	if err := r.parseConfig(config); err != nil {
 		return err
 	}
@@ -179,6 +190,40 @@ func (r *caseRunner) run() error {
 
 	log.Info().Str("testcase", config.Name()).Msg("run testcase end")
 	return nil
+}
+
+func initPlugin(path string) (plugin common.Plugin, err error) {
+	plugin, err = common.Init(path)
+	if plugin == nil {
+		return
+	}
+
+	// catch Interrupt and SIGTERM signals to ensure plugin quitted
+	c := make(chan os.Signal)
+	signal.Notify(c, os.Interrupt, syscall.SIGTERM)
+	go func() {
+		<-c
+		plugin.Quit()
+		os.Exit(0)
+	}()
+
+	// report event for initializing plugin
+	var pluginType string
+	if _, ok := plugin.(*common.GoPlugin); ok {
+		pluginType = "go"
+	} else {
+		pluginType = "hashicorp"
+	}
+	event := ga.EventTracking{
+		Category: "InitPlugin",
+		Action:   fmt.Sprintf("Init %s plugin", pluginType),
+	}
+	if err != nil {
+		event.Value = 1 // failed
+	}
+	go ga.SendEvent(event)
+
+	return
 }
 
 func (r *caseRunner) runStep(index int, caseConfig *TConfig) (stepResult *stepData, err error) {
@@ -502,12 +547,6 @@ func (r *caseRunner) runStepTestCase(step *TStep) (stepResult *stepData, err err
 
 func (r *caseRunner) parseConfig(config IConfig) error {
 	cfg := config.ToStruct()
-
-	// init plugin
-	err := r.parser.initPlugin(cfg.Path)
-	if err != nil {
-		return err
-	}
 
 	// parse config variables
 	parsedVariables, err := r.parser.parseVariables(cfg.Variables)